+++ PaymentButtons/src/main/java/com/paypal/android/paymentbuttons/PayLaterButton.kt
@@ -1,7 +1,10 @@
 package com.paypal.android.paymentbuttons
 
 import android.content.Context
+import android.content.res.TypedArray
 import android.util.AttributeSet
+import android.view.View
+import androidx.core.content.res.use
 import com.paypal.android.ui.R
 
 
@@ -24,14 +27,27 @@ class PayLaterButton @JvmOverloads constructor(
     context: Context,
     attributeSet: AttributeSet? = null,
     defStyleAttr: Int = 0
-) : PayPalButton(context, attributeSet, defStyleAttr) {
+) : PaymentButton<PayPalButtonColor>(context, attributeSet, defStyleAttr) {
+
+    /**
+     * Updates the color of the Payment Button with the provided [PayPalButtonColor].
+     *
+     * This may update the PayPal wordmark to aid with visibility as well. When updated to GOLD or
+     * WHITE it will be updated to the traditional wordmark. When updated to BLUE or BLACK it will
+     * be updated to the monochrome wordmark.
+     */
+    override var color: PaymentButtonColor = PayPalButtonColor.GOLD
+        set(value) {
+            field = value
+            updateShapeDrawableFillColor(field)
+        }
 
     /**
      * Provides the label of the [PayLaterButton]. This value will always be
      * [PaymentButtonFundingType.PAY_LATER], attempting
      * to set the value to anything else will result in that value being ignored.
      */
-    override var label: PayPalButtonLabel = PayPalButtonLabel.PAY_LATER
+    var label: PayPalButtonLabel = PayPalButtonLabel.PAY_LATER
         set(value) {
             if (value == PayPalButtonLabel.PAY_LATER) {
                 field = value
@@ -49,5 +65,29 @@ class PayLaterButton @JvmOverloads constructor(
 
     init {
         updateLabel(PayPalButtonLabel.PAY_LATER)
+        analyticsService.sendAnalyticsEvent(
+            "payment-button:initialized",
+            orderId = null,
+            buttonType = PaymentButtonFundingType.PAY_LATER.buttonType
+        )
+    }
+
+    private fun updateLabel(updatedLabel: PayPalButtonLabel) {
+        when (updatedLabel.position) {
+            PayPalButtonLabel.Position.START -> {
+                suffixTextVisibility = View.GONE
+                prefixTextVisibility = View.VISIBLE
+                prefixText = updatedLabel.retrieveLabel(context)
+            }
+            PayPalButtonLabel.Position.END -> {
+                prefixTextVisibility = View.GONE
+                suffixTextVisibility = View.VISIBLE
+                suffixText = updatedLabel.retrieveLabel(context)
+            }
+            else -> {
+                prefixTextVisibility = View.GONE
+                suffixTextVisibility = View.GONE
+            }
+        }
     }
 }