+++ PaymentButtons/src/main/java/com/paypal/android/paymentbuttons/PaymentButton.kt
@@ -17,6 +17,9 @@ import com.google.android.material.shape.CutCornerTreatment
 import com.google.android.material.shape.MaterialShapeDrawable
 import com.google.android.material.shape.RoundedCornerTreatment
 import com.google.android.material.shape.ShapeAppearanceModel
+import com.paypal.android.corepayments.CoreConfig
+import com.paypal.android.corepayments.Environment
+import com.paypal.android.corepayments.analytics.AnalyticsService
 import com.paypal.android.ui.R
 
 @Suppress("TooManyFunctions")
@@ -30,6 +33,9 @@ abstract class PaymentButton<C : PaymentButtonColor> @JvmOverloads constructor(
      * Signals that the backing shape has changed and may require a full redraw.
      */
     private var shapeHasChanged = false
+
+    internal val analyticsService: AnalyticsService =
+        AnalyticsService(context, CoreConfig(clientId = "N/A", environment = Environment.LIVE))
 
     private var shapeAppearanceModel: ShapeAppearanceModel = ShapeAppearanceModel()
         set(value) {
@@ -222,6 +228,17 @@ abstract class PaymentButton<C : PaymentButtonColor> @JvmOverloads constructor(
         }
     }
 
+    override fun setOnClickListener(listener: OnClickListener?) {
+        super.setOnClickListener { view ->
+            listener?.onClick(view)
+            analyticsService.sendAnalyticsEvent(
+                "payment-button:tapped",
+                orderId = null,
+                buttonType = fundingType.buttonType
+            )
+        }
+    }
+
     private fun updateSizeFrom(typedArray: TypedArray) {
         val paypalSizeAttribute = typedArray.getInt(
             R.styleable.PaymentButton_payment_button_size,
@@ -297,8 +314,8 @@ abstract class PaymentButton<C : PaymentButtonColor> @JvmOverloads constructor(
     }
 }
 
-internal enum class PaymentButtonFundingType {
-    PAYPAL,
-    PAY_LATER,
-    PAYPAL_CREDIT;
+internal enum class PaymentButtonFundingType(val buttonType: String) {
+    PAYPAL("PayPal"),
+    PAY_LATER("Pay Later"),
+    PAYPAL_CREDIT("Credit");
 }